Separated the file save interfaces and classes into separate files and removed the redundant code
