+++ Vts.Gui.Wpf/FileSystem/ISaveFileDialog.cs
@@ -1,0 +1,10 @@
+﻿namespace Vts.Gui.Wpf.FileSystem
+{
+    public interface ISaveFileDialog
+    {
+        string Filter { get; set; }
+        bool? ShowDialog();
+        string FileName { get; set; }
+        string DefaultExtension { get; set; }
+    }
+}
