+++ Vts.Gui.Wpf/FileSystem/ITextFileService.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.IO;
+
+namespace Vts.Gui.Wpf.FileSystem
+{
+    public interface ITextFileService
+    {
+        public Tuple<FileStream, string> OpenTextFile();
+    }
+}
